requirements.txt
flask==2.3.2
flask-cors==3.0.10
firebase-admin==6.2.0
python-dotenv==1.0.0
gunicorn==20.1.0  # Solo para producción
requests==2.31.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9
urllib3==2.0.7